otsuthr/requirements.txt
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==37.0.1
easyocr==1.5.0
idna==3.3
imageio==2.21.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.8.5
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.2.0
pip==22.1.2
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-bidi==0.4.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.0
setuptools==61.2.0
six==1.16.0
tifffile==2022.8.8
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.3.0
UNKNOWN==0.0.0
urllib3==1.26.11
wheel==0.37.1
